Add a test for `objectOf` props
Summary:
A Flow type like `{[name: string]: number}` describes an object with
arbitrary keys all of whose values are numbers. This is represented as
the React PropType `PropTypes.objectOf(PropTypes.number)`, and should be
supported by this plugin.

The test case uses `objectOf(x)` instead of `objectOf(x.isRequired)`
because the existing behavior for `arrayOf` is analogous. Perhaps this
can be changed separately, as it disagrees with Flow's semantics.

The test currently fails; I changed the snapshot file from the generated
output to what it actually should be. Yarn says:

```
 FAIL  src/__tests__/objectof-test.js
  ● objectOf

    expect(value).toMatchSnapshot()

    Received value does not match stored snapshot 1.

    - Snapshot
    + Received

    @@ -34,12 +34,9 @@
       oneNumber: _propTypes2.default.number.isRequired,
       oneComplex: _propTypes2.default.shape({
         real: _propTypes2.default.number.isRequired,
         imag: _propTypes2.default.number.isRequired
       }).isRequired,
    -  manyNumbers: _propTypes2.default.objectOf(_propTypes2.default.number).isRequired,
    -  manyComplex: _propTypes2.default.objectOf(_propTypes2.default.shape({
    -    real: _propTypes2.default.number.isRequired,
    -    imag: _propTypes2.default.number.isRequired
    -  })).isRequired
    +  manyNumbers: _propTypes2.default.shape({}).isRequired,
    +  manyComplex: _propTypes2.default.shape({}).isRequired
     };
     exports.default = Foo;"

      at Object.<anonymous> (src/__tests__/objectof-test.js:27:15)
      at Promise.resolve.then.el (node_modules/p-map/index.js:46:16)
      at process._tickCallback (internal/process/next_tick.js:109:7)
```

wchargin-branch: objectof-test